Updated many files with bug fixes after submission.

+++ course1/week1_programming_challenges/2_maximum_pairwise_product/max_pairwise_product.cpp
@@ -5,11 +5,11 @@
 using namespace std;
 
 
-int MaxPairwiseProduct(const vector<long long>& numbers) {
+unsigned long long MaxPairwiseProduct(const vector<unsigned long long>& numbers) {
     int n = numbers.size();
 
-    long long num1 = max(numbers[0], numbers[1]);
-    long long num2 = min(numbers[0], numbers[1]);
+    unsigned long long num1 = max(numbers[0], numbers[1]);
+    unsigned long long num2 = min(numbers[0], numbers[1]);
 
     for (int i = 2; i < n; ++i) {
         if (numbers[i] > num1) {
@@ -25,12 +25,12 @@ int MaxPairwiseProduct(const vector<long long>& numbers) {
 int main() {
     int n;
     cin >> n;
-    vector<long long> numbers(n);
+    vector<unsigned long long> numbers(n);
     for (int i = 0; i < n; ++i) {
         cin >> numbers[i];
     }
 
-    long long result = MaxPairwiseProduct(numbers);
+    unsigned long long result = MaxPairwiseProduct(numbers);
     cout << result << "\n";
     return 0;
 }
